backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
alembic==1.17.0
asyncpg==0.30.0
black==25.9.0
fastapi==0.118.3
flake8==7.3.0
flake8-bugbear==24.12.12
flake8-comprehensions==3.17.0
flake8-print==5.0.0
google-genai==1.43.0
hf-xet==1.1.10
isort==7.0.0
langchain==0.3.27
langchain-community==0.3.31
langchain-huggingface==0.3.1
langchain-postgres==0.0.15
langgraph==0.6.10
langgraph-checkpoint-postgres==2.0.25
openpyxl==3.1.5
pandas==2.3.3
psycopg2-binary==2.9.11
pydantic==2.12.0
pydantic-settings==2.11.0
pypdf==6.1.1
python-docx==1.2.0
python-multipart==0.0.20
semantic-router==0.1.11
sentence-transformers==5.1.1
strawberry-graphql[fastapi]==0.283.3
sqlalchemy==2.0.44
torch==2.8.0
unstructured==0.18.15
uvicorn==0.37.0
